backend/requirements.txt
itsdangerous==2.2.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.17
pydantic-settings==2.6.1
email-validator==2.2.0
httpx==0.28.1
authlib==1.4.0
pytest==8.3.4
